requirements.txt
# -------------------------------
# Conversational Health Assistant
# LangChain Symptom Checker Project
# -------------------------------
python-dotenv>=1.0.0

# Core LLM + Chain libraries
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0

# Vector stores for document search
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Embeddings and token management
tiktoken>=0.7.0

# PDF document handling
pypdf>=3.15.0

# Utility for generating local dataset PDFs
fpdf>=1.7.2

# Memory, chains, and toolkits dependencies
numpy>=1.24.0
pandas>=2.0.0

